fix(windows): Link statically with winpthreads
This closes #13, which only affected i686 installers. That explains why
the issue was not as widely reported as expected.

`-static` might be overkill, it's possible to laser-focus winpthreads
with `-Wl,-Bstatic -lpthread -Wl,-Bdynamic`, but it's unclear how that
would interact with the `-lpthread` the Go compile passes by itself,
and how spaces would be escaped since the full command is
`-ldflags="-extldflags='GNU ld flags here'"`.

In any case we never want itch-setup to depend on DLLs that don't ship
with Windows itself, since it's meant to be a self-contained executable.